Initial commit - Add ability to redirect to a splash screen URL
Checks for the presence of a cookie with specific values when activated,
and redirect the user to a configurable URL when it is not found, or not
with an accepted value.

To be used to display a splash screen to users upon the first visit.
It's the responsability of the splash page at the redirected URL to set
the proper cookie value before sending users back to the LMS.